+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/Common/Content/ContentValueDtoValidator.cs
@@ -1,5 +1,5 @@
-using Digdir.Domain.Dialogporten.Application.Common;
 using Digdir.Domain.Dialogporten.Application.Common.Authorization;
+using System.Diagnostics.CodeAnalysis;
 using Digdir.Domain.Dialogporten.Application.Common.Extensions;
 using Digdir.Domain.Dialogporten.Application.Common.Extensions.FluentValidation;
 using Digdir.Domain.Dialogporten.Application.Externals.Presentation;
@@ -12,11 +12,7 @@ using FluentValidation;
 namespace Digdir.Domain.Dialogporten.Application.Features.V1.Common.Content;
 
 // DialogContentValueDtoValidator has constructor parameter input, and can't be registered in DI assembly scan
-// This interface is used to ignore the class when scanning for validators
-// The validator is manually created in the Create and Update validators
-
-// DialogContentValueDtoValidator has constructor parameter input, and can't be registered in DI assembly scan
-// This interface is used to ignore the class when scanning for validators
+// IIgnoreOnAssemblyScan is used to ignore the class when scanning for validators
 // The validator is manually created in the Create and Update validators
 internal sealed class ContentValueDtoValidator : AbstractValidator<ContentValueDto>, IIgnoreOnAssemblyScan
 {
@@ -71,22 +67,15 @@ internal sealed class ContentValueDtoValidator : AbstractValidator<ContentValueD
             .SetValidator(_ => new LocalizationDtosValidator(contentType.MaxLength));
     }
 
+    [SuppressMessage("Style", "IDE0072:Add missing cases")]
     private static string[] GetAllowedMediaTypes(DialogContentType contentType, IUser? user)
-    {
-        if (user == null)
+        => contentType.Id switch
         {
-            return contentType.AllowedMediaTypes;
-        }
-
-        if (contentType.Id != DialogContentType.Values.AdditionalInfo)
-        {
-            return contentType.AllowedMediaTypes;
-        }
-
-        var allowHtmlSupport = user.GetPrincipal().HasScope(Constants.LegacyHtmlScope);
-
-        return allowHtmlSupport
-            ? contentType.AllowedMediaTypes.Append(LegacyHtmlMediaType).ToArray()
-            : contentType.AllowedMediaTypes;
-    }
+            DialogContentType.Values.AdditionalInfo when UserHasLegacyHtmlScope(user)
+                => contentType.AllowedMediaTypes.Append(MediaTypes.LegacyHtml).ToArray(),
+            DialogContentType.Values.MainContentReference when UserHasLegacyHtmlScope(user)
+                => contentType.AllowedMediaTypes.Append(MediaTypes.LegacyEmbeddableHtml).ToArray(),
+            _ => contentType.AllowedMediaTypes
+        };
+    private static bool UserHasLegacyHtmlScope(IUser? user) => user is not null && user.GetPrincipal().HasScope(Constants.LegacyHtmlScope);
 }